Rename connector package cdc=>debezium. Update doc references Fix code style on DebeziumIOMySqlConnectorIT
