+++ sdks/java/io/debezium/build.gradle
@@ -19,7 +19,7 @@ import groovy.json.JsonOutput
 
 plugins { id 'org.apache.beam.module' }
 applyJavaNature(
-        automaticModuleName: 'org.apache.beam.sdk.io.cdc',
+        automaticModuleName: 'org.apache.beam.sdk.io.debezium',
         mavenRepositories: [
                 [id: 'io.confluent', url: 'https://packages.confluent.io/maven/']
         ],
@@ -28,8 +28,8 @@ applyJavaNature(
 )
 provideIntegrationTestingDependencies()
 
-description = "Apache Beam :: SDKs :: Java :: IO :: CDC"
-ext.summary = "Library to work with CDC data."
+description = "Apache Beam :: SDKs :: Java :: IO :: Debezium"
+ext.summary = "Library to work with Debezium data."
 
 dependencies {
     compile library.java.vendored_guava_26_0_jre
