+++ sdks/java/io/debezium/src/main/java/org/apache/beam/io/debezium/SourceRecordMapper.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.io.cdc;
+package org.apache.beam.io.debezium;
 
 import java.io.Serializable;
 import org.apache.kafka.connect.source.SourceRecord;
