+++ sdks/java/io/debezium/src/test/java/org/apache/beam/io/debezium/DebeziumIOTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.io.cdc;
+package org.apache.beam.io.debezium;
 
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
@@ -25,7 +25,7 @@ import io.debezium.connector.mysql.MySqlConnector;
 import io.debezium.connector.mysql.MySqlConnectorConfig;
 import java.io.Serializable;
 import java.util.Map;
-import org.apache.beam.io.cdc.DebeziumIO.ConnectorConfiguration;
+import org.apache.beam.io.debezium.DebeziumIO.ConnectorConfiguration;
 import org.apache.kafka.common.config.ConfigValue;
 import org.junit.Test;
 import org.junit.runner.RunWith;
