+++ settings.gradle
@@ -129,7 +129,7 @@ include ":sdks:java:io:amazon-web-services2"
 include ":sdks:java:io:amqp"
 include ":sdks:java:io:azure"
 include ":sdks:java:io:cassandra"
-include ":sdks:java:io:cdc"
+include ":sdks:java:io:debezium"
 include ":sdks:java:io:clickhouse"
 include ":sdks:java:io:common"
 include ":sdks:java:io:contextualtextio"
